% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat1000}
\alias{dat1000}
\title{Bootstrap results for m4id = 9057756 using 1000 replicates}
\format{A data.table with 1000 rows and 59 columns}
\usage{
dat1000
}
\description{
A dataset containing the bootstrapped results of m4id = 9057756 found in the
erl3data dataset. Used as an example for plotting in the vignette, which
shows the code used to generate the data.
}
\keyword{datasets}

