% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erl5data}
\alias{erl5data}
\title{Fit parameters, model selection, and hit calls from 200 ER assay curves}
\format{A data.table with 200 rows and 67 variables}
\source{
\url{https://www.epa.gov/chemical-research/toxicity-forecaster-toxcasttm-data/}
}
\usage{
erl5data
}
\description{
A dataset containing the ToxCast level 5 data corresponding to the m4ids in
\code{\link{erl3data}}. This data is provided so that the cutoff values for
these assays can be used in \code{\link{toxbootHitParamCI}} to determine
winning model and hit call selection for bootstrap results. This dataset is
also useful to compare min/med/max bootstrap results to the point estimate in
the ToxCast pipeline value.
}
\keyword{datasets}

