% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{toxbootGetOpts}
\alias{toxbootGetOpts}
\alias{toxbootSetOpts}
\title{DEFUNCT}
\usage{
toxbootGetOpts()
}
\description{
These functions were renamed and now have diverged. They are no longer
compatible and will not run.
}
\details{
The following functions are defunct and will be removed; use
the replacements indicated below:
  \itemize{
    \item{toxbootGetOpts: \code{\link{toxbootConfList}}}
    \item{toxbootSetOpts: \code{\link{toxbootConf}}}
  }
}
