% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootMongoFuncs.R
\name{toxbootConnectMongo}
\alias{toxbootConnectMongo}
\title{Setup connection to toxboot MongoDB}
\usage{
toxbootConnectMongo()
}
\value{
mongo, a mongo connection object.
}
\description{
\code{toxbootConnectMongo} imports toxboot options and creates the connection to the mongoDB.
}

