% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootMongoFuncs.R
\name{toxbootGetMongoFields}
\alias{toxbootGetMongoFields}
\title{Query mongoDB and get requested fields}
\usage{
toxbootGetMongoFields(fields, ...)
}
\arguments{
\item{fields}{String, a vector of fields to return from the database. Passed
to \code{toxbootProjectionBuild}}

\item{...}{parameters to query on, passed to \code{toxbootQueryBuild}.
Format is query_field = query_values}
}
\value{
data table of requested fields
}
\description{
\code{toxbootGetMongoFields} will query the mongo database for matching
bootstrap type(s), assay aeid(s), and sample id(s). It will then return a
data.table with the requested fields.
}
\details{
Longer description here
}
\examples{
\dontrun{
toxbootGetMongoFields(boot_type = "smooth", aeid = 1508L,
fields = c("hill_ga", "gnls_ga"))
}

}
