% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootGetOpts.R
\name{toxbootGetOpts}
\alias{toxbootGetOpts}
\title{DEPRECATED}
\usage{
toxbootGetOpts(show.pass = FALSE)
}
\arguments{
\item{show.pass}{Logical, should the password be returned}
}
\description{
\code{toxbootGetOtps} is deprecated. See \code{toxbootConfList}.
}

