% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootMonitorFuncs.R
\name{toxbootMongoCount}
\alias{toxbootMongoCount}
\title{Get number of documents that match query.}
\usage{
toxbootMongoCount(mongo, query = NULL)
}
\arguments{
\item{mongo, }{a mongo connection object returned from
\code{toxbootConnectMongo()}}

\item{query, }{a mongo.bson object returned from \code{toxbootQueryBuild}. By
default query will match all documents in the collection.}
}
\value{
num_results, an integer equal to the number of documents that match the query.
}
\description{
\code{toxbootMongoCount} returns the number of objects that match the query.
}
