% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootMongoFuncs.R
\name{toxbootProjectionBuild}
\alias{toxbootProjectionBuild}
\title{Build projection for toxboot MongoDB}
\usage{
toxbootProjectionBuild(fields, id_val = 0L)
}
\arguments{
\item{fields}{string vector, fields to return.}

\item{id_val}{integer 0L or 1L. By default 0L, which will supress Object_ID
from the projection. This unique ID is often not needed. Can be set to 1L
to include in projection.}
}
\value{
projection, a named list where names are the fields to return in the
  projection and the values are 0L (don't return) or 1L (return).
}
\description{
\code{toxbootProjectionBuild} builds a list that determined which fields
are returned.
}

