% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootMongoFuncs.R
\name{toxbootQueryBuild}
\alias{toxbootQueryBuild}
\title{Build query for toxboot MongoDB}
\usage{
toxbootQueryBuild(...)
}
\arguments{
\item{...}{parameters to query on. Format is query_field = query_values}
}
\value{
query, a mongo.bson object used for a query
}
\description{
\code{toxbootQueryBuild} imports toxboot query from list of input parameters.
}

