% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootQueryToxCast.R
\name{toxbootQueryToxCast}
\alias{toxbootQueryToxCast}
\title{Query ToxCast and format data to pass to \code{toxboot}}
\usage{
toxbootQueryToxCast(aeids)
}
\arguments{
\item{aeids}{integer vector contains which aeids to query. Values will be
passed to tcplLoadData.}
}
\value{
dat, a data.table corresponding to the level 3 data for the aeids
  with columns added for \code{m4id} and \code{bmad}.
}
\description{
\code{toxbootQueryToxCast} passes .
}
\details{
Because \code{bmad} is calculated for all chemicals in the assay,
  the entire aeid must be queried. After calculation of \code{bmad} chemicals
  can be further subset prior to running toxboot.

  The assumption is that \code{\link{tcplConf}} has been setup to pull data
  from the correct database.
}
