% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootReplicates.R
\name{toxbootReplicates}
\alias{toxbootReplicates}
\title{Generates bootstrap samples of dose response data}
\usage{
toxbootReplicates(datchemval, boot_method = "smooth", replicates = 100)
}
\arguments{
\item{datchemval}{data.table, columns logc, resp, bmad}

\item{boot_method}{string, specifies method. Default is 'smooth'.}

\item{replicates}{number of bootstrap samples. Default 100}
}
\description{
\code{toxbootReplicates} takes input dose response data and returns
a matrix of sampled dose response values.
}
\details{
Accepted methods are "case", "smooth", "residuals_hill", "wild_hill",
"residuals_gnls", "wild_gnls", "watt_normal", "watt_student_df5"
}
\seealso{
\code{\link{toxboot}}
}
