% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootSetOpts.R
\name{toxbootSetOpts}
\alias{toxbootSetOpts}
\title{DEPRECATED}
\usage{
toxbootSetOpts(mongo_host = NULL, DBNS = NULL, user = NULL, pass = NULL,
  db = NULL)
}
\arguments{
\item{mongo_host}{Character of length 1, database IP address}

\item{DBNS}{Character of length 1, the collection on the database to read
from and write to}

\item{user}{Character of length 1, username to authenticate}

\item{pass}{Character of length 1, password that corresponds to username}

\item{db}{Character of length 1, database where the username and pass are
authenicated. Can be an IP address or a url.}
}
\description{
\code{toxbootSetOtps} is deprecated. See \code{toxbootConf}.
}

