% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxbootMySQLSetup.R
\name{toxbootMysqlCreateTable}
\alias{toxbootMysqlCreateTable}
\title{Function to create properly formatted toxboot table in MySQL}
\usage{
toxbootMysqlCreateTable(table_name = NULL)
}
\arguments{
\item{table_name}{string length 1, the name of the MySQL table format and
create}
}
\description{
\code{toxbootMysqlCreateTable} creates table with correct column types which
will be written to by \code{toxboot} and \code{toxbootmc}.
}
\details{
Caution. This function will delete the current toxboot table if it
  exists. Still need to add some safeguards.
}
